+++ src/components/download/DownloadSettings.js
@@ -6,18 +6,13 @@ import { setDownloadConfig } from '../../actions/download.js'
 import { standardizeFilename } from '../../util/dataDownload.js'
 import { downloadMapImage, downloadSupport } from '../../util/export-image.js'
 import { getSplitViewLayer } from '../../util/helpers.js'
-import history from '../../util/history.js'
+import { closeDownloadMode } from '../../util/history.js'
 import { getMapName } from '../app/FileMenu.js'
 import Drawer from '../core/Drawer.js'
 import { Checkbox, Help } from '../core/index.js'
 import LegendLayers from './LegendLayers.js'
 import NorthArrowPosition from './NorthArrowPosition.js'
 import styles from './styles/DownloadSettings.module.css'
-
-const closeDownloadMode = () => {
-    const rootPath = history.location.pathname.split('/download')[0]
-    history.push(rootPath, { isDownloadClosing: true })
-}
 
 const DownloadSettings = () => {
     const [error, setError] = useState(null)
